finish norme parsing and divid by files, next step error and memory
