+++ .mk/sources.mk
@@ -14,14 +14,19 @@ SRC = \
     $(SRC_DIR)cd.c \
     $(SRC_DIR)utils.c \
     \
+    $(PARSING_DIR)lexer.c \
+    $(PARSING_DIR)parser.c \
     $(PARSING_DIR)buffer_dynamic.c \
     $(PARSING_DIR)expand.c \
     $(PARSING_DIR)generate_temp_filename.c \
     $(PARSING_DIR)group_tokens.c \
     $(PARSING_DIR)heredoc_handle.c \
-    $(PARSING_DIR)lexer_utils.c \
-    $(PARSING_DIR)lexer.c \
-    $(PARSING_DIR)parser.c \
+    $(PARSING_DIR)lexer_flush.c \
+    $(PARSING_DIR)lexer_normal.c \
+    $(PARSING_DIR)lexer_quote.c \
+    $(PARSING_DIR)segment.c \
+    $(PARSING_DIR)token_list.c \
+    $(PARSING_DIR)create_token.c \
     $(PARSING_DIR)print_var.c \
     \
     $(PARSING_UTILS)ft_isalnum.c \
@@ -36,6 +41,7 @@ SRC = \
     $(PARSING_UTILS)ft_itoa.c \
     $(PARSING_UTILS)get_int_size.c \
     $(PARSING_UTILS)append_all.c \
+    $(PARSING_UTILS)is_fonction.c \
     $(PARSING_UTILS)ft_strlen.c
 
 HEADERS = \
